Add support for submeshing
The scene is no longer assumed to be a single mesh. The keys `vertices`,
`uvs`, `normals`, `faces`, `skinWeights`, and `skinIndices` have all
been moved to a separate array of objects under the key `meshes`. The
`bakeAnimations` option was also supported as it is unused by the
current engines/file formats

WIP